YEL-7394 Add inactive notprod instance ip to exclusion
Add a function to read ssl.conf and extract the ip address
of the inactive notprod ec2 instance.
Add inactive notprod instance ip address to list of excluded
ec2 instances.
On daily startup process, this instance will be
ignored and will remain in a stopped state, without the
need for manual intervention.
This assumes that the line containing the inactive instance
ip in ssl.conf is commented with a `#` and not by some other
pattern.