YEL-7394 Fix regular expression
Fix the regular expression to properly extract the inactive
instance ip address. If it appears on a commented out line,
i.e. it begins with a '#' character, then this is the ip
that should be added to the exclusion list. Allowed for
additional whitespace characters surrounding the address to
be more forgiving if somebody adccidentally adds a space or
a tab to the config for example.